+++ src/effects/builtin/dynamics/limiter/limiterviewmodel.cpp
@@ -14,31 +14,4 @@ LimiterViewModel::LimiterViewModel(QObject* parent)
 void LimiterViewModel::doReload()
 {
 }
-
-LimiterSettingModel::LimiterSettingModel(QObject* parent)
-    : EffectSettingModelImpl<LimiterEffect>(parent, [this](const LimiterEffect& effect) {
-    if (m_paramId == "thresholdDb") {
-        return effect.thresholdDb;
-    } else if (m_paramId == "makeupTargetDb") {
-        return effect.makeupTargetDb;
-    } else if (m_paramId == "kneeWidthDb") {
-        return effect.kneeWidthDb;
-    } else if (m_paramId == "lookaheadMs") {
-        return effect.lookaheadMs;
-    } else if (m_paramId == "releaseMs") {
-        return effect.releaseMs;
-    } else if (m_paramId == "showInput") {
-        return effect.showInput;
-    } else if (m_paramId == "showOutput") {
-        return effect.showOutput;
-    } else if (m_paramId == "showActual") {
-        return effect.showActual;
-    } else if (m_paramId == "showTarget") {
-        return effect.showTarget;
-    } else {
-        LOGE("Unknown parameter ID: ") << m_paramId;
-        assert(false);
-        return effect.thresholdDb;
-    }
-}) {}
 }
