+++ src/effects/builtin/internal/builtineffectsrepository.cpp
@@ -30,8 +30,10 @@
 #include "dynamics/timeline/dynamicsplaystatemodel.h"
 #include "dynamics/compressor/compressoreffect.h"
 #include "dynamics/compressor/compressorviewmodel.h"
+#include "dynamics/compressor/compressorsettingmodel.h"
 #include "dynamics/limiter/limitereffect.h"
 #include "dynamics/limiter/limiterviewmodel.h"
+#include "dynamics/limiter/limitersettingmodel.h"
 #include "normalize/normalizeeffect.h"
 #include "normalize/normalizeviewmodel.h"
 #include "tonegen/chirpeffect.h"
